+++ db/schema.rb
@@ -30,16 +30,16 @@ ActiveRecord::Schema.define(version: 20150512122110) do
     t.text     "selected_option"
     t.text     "correct_option"
     t.text     "option_status"
-    t.integer  "option_score",    default: 0
-    t.datetime "created_at",                  null: false
-    t.datetime "updated_at",                  null: false
+    t.float    "option_score",    default: 0.0
+    t.datetime "created_at",                    null: false
+    t.datetime "updated_at",                    null: false
   end
 
   create_table "users", force: :cascade do |t|
     t.string   "name",                   default: "",   null: false
     t.string   "email",                  default: "",   null: false
     t.string   "category",               default: "jr", null: false
-    t.integer  "total_score",            default: 0,    null: false
+    t.float    "total_score",            default: 0.0,  null: false
     t.integer  "time_spent",             default: 0,    null: false
     t.string   "encrypted_password",     default: "",   null: false
     t.string   "reset_password_token"
